Refactored compute_score so problem definition is defined in a compact way. Also added verbose option when computing score.
